requirements.txt
beautifulsoup4==4.12.2
selenium==4.10.0
spotipy==2.22.1
python-dotenv==1.0.0
requests==2.31.0